pyproject.toml
[project]
name = "airportdirectory"
version = "0.1.0"
description = ""
authors = [
    {name = "Станислав Рубцов",email = "frex@mail.ru"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pydantic-async-validation (>=0.3.0,<0.4.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "fastapi-pagination (>=0.12.34,<0.13.0)",
    "httpx (==0.27.0)",
    "authlib (>=1.5.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
mypy = "^1.15.0"
isort = "^6.0.1"
flake8 = "^7.1.2"

